Split tests to make easier to use
