+++ hostinfo/host/api_urls.py
@@ -1,5 +1,5 @@
 """ API Url handler """
-from django.urls import path
+from django.urls import path, re_path
 
 from .rest_views import (
     HostAliasRest,
@@ -18,28 +18,31 @@ linkspec = "(<int:linkpk>|<str:tagname>)"
 
 urlpatterns = [
     path("alias/", AliasList),
-    path(
+    re_path(
         f"host/{hostspec}/alias/{aliasspec}/", HostAliasRest, name="hostaliasrest"
     ),
-    path(f"host/{hostspec}/alias/", HostAliasRest, name="hostaliasrest"),
-    path(
+    re_path(f"host/{hostspec}/alias/", HostAliasRest, name="hostaliasrest"),
+    re_path(
         f"host/{hostspec}/key/{kvalspec}/<str:value>/",
         HostKeyRest,
         name="hostkeyrest",
     ),
-    path(f"host/{hostspec}/key/{kvalspec}/", HostKeyRest, name="hostkeyrest"),
-    path(
+    re_path(f"host/{hostspec}/key/{kvalspec}/", HostKeyRest, name="hostkeyrest"),
+    re_path(
         f"host/{hostspec}/link/{linkspec}/<str:url>",
         HostLinkRest,
         name="hostlinkrest",
     ),
-    path(f"host/{hostspec}/link/{linkspec}/", HostLinkRest, name="hostlinkrest"),
-    path(f"host/{hostspec}/", HostDetail, name="resthost"),
+    re_path(f"host/{hostspec}/link/{linkspec}/", HostLinkRest, name="hostlinkrest"),
+    path("host/<int:hostpk>/", HostDetail, name="resthost"),
+    path("host/<str:hostname>/", HostDetail, name="resthost"),
     path("host/", HostList),
-    path(f"key/{akeyspec}", KeyDetail, name="restakey"),
+    path("key/<int:akeypk>", KeyDetail, name="restakey"),
+    path("key/<str:akey>", KeyDetail, name="restakey"),
     path("kval/<int:pk>/", KValDetail, name="restkval"),
     path("query/<str:query>/", HostQuery),
-    path(f"keylist/{akeyspec}/<str:query>", KeyListRest),
+    path("keylist/<int:akeypk>/<str:query>", KeyListRest),
+    path("keylist/<str:akey>/<str:query>", KeyListRest),
 ]
 
 # EOF
