+++ hostinfo/report/tests/tests.py
