+++ hostinfo/runtests
@@ -1,7 +1,7 @@
 #!/bin/sh
 
-export PYTHONPATH=hostinfo
+export PYTHONPATH=hostinfo:.
 TEST="./manage.py test host.tests --settings hostinfo.settings"
-coverage run --include 'host/*,hostinfo/*' ${TEST} && coverage report -m
+coverage run --include 'hostinfo' ${TEST} && coverage report -m
 
 #EOF
